# openapi.yaml
openapi: 3.0.3
info:
  title: AI CLI Manager Daemon API
  version: 0.1.0
  description: |
    REST API for managing multiple AI CLI instances on a Windows host daemon.

    Notes:
      - This spec covers REST endpoints under `/api/v1`.
      - WebSocket endpoints are documented under `/ws/v1/*` using a conventional `101 Switching Protocols`
        response. OpenAPI does not formally model WebSocket message flows; see `docs/API.md` for the
        detailed WS message protocol.
servers:
  - url: http://127.0.0.1:8765
    description: Local daemon (default)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/instances:
    get:
      tags: [Instances]
      summary: List instances
      parameters:
        - name: include_runtime
          in: query
          description: Include runtime fields (status/pid/metrics) if available.
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: List of instances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Instances]
      summary: Create a new instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstanceCreateRequest"
      responses:
        "201":
          description: Instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/instances/{id}/config:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/InstanceId"
    get:
      tags: [Config]
      summary: Get instance config view
      parameters:
        - name: include_content
          in: query
          description: Include config content (may be large). For mode=path, daemon may read from file.
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Instance config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Config]
      summary: Update instance config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstanceConfigUpdateRequest"
      responses:
        "200":
          description: Updated config view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    InstanceId:
      type: string
      format: uuid
      example: "8d8a4d3c-3d50-4d87-9a7f-4df9dbf9c9f8"
    Timestamp:
      type: string
      format: date-time
      example: "2026-03-01T12:34:56Z"
...
    RestartPolicy:
      type: string
      enum: [never, on-failure, always]
    ConfigMode:
      type: string
      enum: [none, path, inline]
    InstanceStatus:
      type: string
      enum: [stopped, starting, running, stopping, exited, error]
    InstanceBase:
      type: object
      required: [name, enabled, command, args, env, use_pty, config_mode, restart_policy, auto_start]
      properties:
        name:
          type: string
          example: "Aider (repo-1)"
        enabled:
          type: boolean
          default: true
        command:
          type: string
          example: "aider"
        args:
          type: array
          items:
            type: string
          example: ["--model", "gpt-4.1", "--verbose"]
        cwd:
          type: string
          nullable: true
          example: "C:\\work\\repo-1"
        env:
          type: object
          additionalProperties:
            type: string
          example:
            OPENAI_API_KEY: "********"
        use_pty:
          type: boolean
          default: true
        config_mode:
          $ref: "#/components/schemas/ConfigMode"
        config_path:
          type: string
          nullable: true
          description: Only meaningful when config_mode=path
          example: "C:\\work\\repo-1\\.aider.conf.yml"
        config_filename:
          type: string
          nullable: true
          description: Only meaningful when config_mode=inline
          example: "config.yaml"
        config_content:
          type: string
          nullable: true
          description: Only meaningful when config_mode=inline
          example: "model: gpt-4.1\n"
        restart_policy:
          $ref: "#/components/schemas/RestartPolicy"
        auto_start:
          type: boolean
          default: false
    InstanceCreateRequest:
      allOf:
        - $ref: "#/components/schemas/InstanceBase"
...
    InstanceRuntime:
      type: object
      required: [status, clients_attached]
      properties:
        status:
          $ref: "#/components/schemas/InstanceStatus"
        pid:
          type: integer
          nullable: true
          example: 12345
        started_at:
          $ref: "#/components/schemas/Timestamp"
          nullable: true
        exit_code:
          type: integer
          nullable: true
          example: 0
        cpu_percent:
          type: number
          format: float
          nullable: true
          example: 2.5
        mem_bytes:
          type: integer
          format: int64
          nullable: true
          example: 104857600
        clients_attached:
          type: integer
          example: 1
    Instance:
      type: object
      required:
        - id
        - created_at
        - updated_at
        - name
        - enabled
        - command
        - args
        - env
        - use_pty
        - config_mode
        - restart_policy
        - auto_start
      properties:
        id:
          $ref: "#/components/schemas/InstanceId"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        name:
          type: string
        enabled:
          type: boolean
        command:
          type: string
        args:
          type: array
          items:
            type: string
        cwd:
          type: string
          nullable: true
        env:
          type: object
          additionalProperties:
            type: string
        use_pty:
          type: boolean
        config_mode:
          $ref: "#/components/schemas/ConfigMode"
        config_path:
          type: string
          nullable: true
        config_filename:
          type: string
          nullable: true
        config_content:
          type: string
          nullable: true
        restart_policy:
          $ref: "#/components/schemas/RestartPolicy"
        auto_start:
          type: boolean
        runtime:
          $ref: "#/components/schemas/InstanceRuntime"
          nullable: true
    InstanceResponse:
      type: object
      required: [instance]
      properties:
        instance:
          $ref: "#/components/schemas/Instance"
    InstanceListResponse:
      type: object
      required: [instances]
      properties:
        instances:
          type: array
          items:
            $ref: "#/components/schemas/Instance"
...
    InstanceConfig:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/ConfigMode"
        path:
          type: string
          nullable: true
          description: Present when mode=path
        filename:
          type: string
          nullable: true
          description: Present when mode=inline
        content:
          type: string
          nullable: true
          description: Present when include_content=true
    InstanceConfigResponse:
      type: object
      required: [id, config]
      properties:
        id:
          $ref: "#/components/schemas/InstanceId"
        config:
          $ref: "#/components/schemas/InstanceConfig"
    InstanceConfigUpdateRequest:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/ConfigMode"
        path:
          type: string
          nullable: true
        filename:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "bad_request"
            message:
              type: string
              example: "Invalid args field; expected JSON array."
            details:
              type: object
              additionalProperties: true
              nullable: true